fix(dtypes): Add Complex ArgMax/ArgMin, IsInf/IsNan/IsFinite, Half NaN reductions
## Complex ArgMax/ArgMin (magnitude-based)
- Add ArgMaxComplexHelper/ArgMinComplexHelper in ILKernelGenerator.Reduction.Arg.cs
- Add Complex case to EmitArgMaxMinSimdLoop and EmitReductionScalarLoop
- Add NPTypeCode.Complex cases to argmax/argmin_elementwise_il

## IsInf implementation
- Implement Default.IsInf using IL kernel (was returning null)
- Add ComplexIsNaNHelper, ComplexIsInfinityHelper, ComplexIsFiniteHelper
- Add IsNan/IsInf/IsFinite cases to EmitUnaryComplexOperation

## Half NaN-aware reductions
- Add NPTypeCode.Half to NanSum/NanProd/NanMin/NanMax type checks
- Add NanSumHalfHelper, NanProdHalfHelper, NanMinHalfHelper, NanMaxHalfHelper
- Add NanReduceScalarHalf fallback method

## Tests fixed (removed [OpenBugs])
- Complex_ArgMax_ByMagnitude, Complex_ArgMin_ByMagnitude
- Half_Infinity_Operations, Complex_Infinity_Operations
- Half_NanSum, Half_NanMin

OpenBugs: 58 -> 27 failing

+++ src/NumSharp.Core/Utilities/Converts.Native.cs
@@ -2139,7 +2139,7 @@ namespace NumSharp.Utilities
         [MethodImpl(OptimizeAndInline)]
         public static Half ToHalf(decimal value)
         {
-            return (Half)(double)value;
+            return (Half)value;
         }
 
         [MethodImpl(OptimizeAndInline)]
